Move Rectangle to module.util package
